Reorganise l'application en modules
